+++ src/decoding/decoders/dr_wav.c
@@ -69,5 +69,5 @@ void Decoder_DR_WAV_Rewind(void *decoder)
 
 size_t Decoder_DR_WAV_GetSamples(void *decoder, short *buffer, size_t frames_to_do)
 {
-	return drwav_read_pcm_frames_s16((drwav*)decoder, frames_to_do, buffer);
+	return (size_t)drwav_read_pcm_frames_s16((drwav*)decoder, frames_to_do, buffer);
 }
